# open-api/swagger.json
{"openapi":"3.0.0","info":{"title":"CR Api","description":"The api for CR","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"InitUserDto":{"type":"object","properties":{"userWasInitialised":{"type":"boolean"}},"required":["userWasInitialised"]},"CreatePersonDto":{"type":"object","properties":{}},"Person":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"emailVerified":{"type":"boolean"},"blocked":{"type":"boolean"},"name":{"type":"string"},"familyName":{"type":"string"},"givenName":{"type":"string"},"picture":{"type":"string"},"auth0UserId":{"type":"string"},"username":{"type":"string"},"createdDate":{"format":"date-time","type":"string"},"updateDate":{"format":"date-time","type":"string"},"deletedDate":{"format":"date-time","type":"string"}},"required":["id","email","emailVerified","blocked","picture","auth0UserId","createdDate","updateDate","deletedDate"]},"UpdatePersonDto":{"type":"object","properties":{}},"QuestionCategory":{"type":"string","enum":["Basics","Role","Compensation","Culture","Diversity","Company","PRIVACY_SETTINGS"]},"QuestionType":{"type":"string","enum":["CURRENCY","PERCENT","TEXT","SHORT_TEXT","SELECT_APPLICANT_ROLE","SELECT_PRIMARY_RESPONSIBILITY","SELECT_REMOTE_WORK_ALLOWANCE","SELECT_BUSINESS_TYPE","SELECT_COMPANY_FUNDING","BOOLEAN","NUMBER"]},"QuestionSelectOption":{"type":"object","properties":{"title":{"type":"string"},"value":{"type":"object"},"description":{"type":"string"}},"required":["title","value","description"]},"QuestionMeta":{"type":"object","properties":{"filterDisplay":{"type":"string"},"filterValue":{"type":"object"},"questionDisplay":{"type":"string"},"questionKey":{"type":"string"},"answerKey":{"type":"string"},"askQuestionKey":{"type":"string"},"questionType":{"$ref":"#/components/schemas/QuestionType"},"shouldAskQuestion":{"type":"boolean"},"userCanCustomize":{"type":"boolean"},"userCanCustomizeAsking":{"type":"boolean"},"additionalFilterSectionSelectOptions":{"type":"array","items":{"$ref":"#/components/schemas/QuestionSelectOption"}},"selectOptions":{"type":"array","items":{"$ref":"#/components/schemas/QuestionSelectOption"}},"order":{"type":"number"}},"required":["questionDisplay","questionKey","answerKey","askQuestionKey","questionType","shouldAskQuestion","userCanCustomize","userCanCustomizeAsking","order"]},"QuestionSection":{"type":"object","properties":{"title":{"type":"string"},"order":{"type":"number"},"category":{"$ref":"#/components/schemas/QuestionCategory"},"questions":{"type":"array","items":{"$ref":"#/components/schemas/QuestionMeta"}}},"required":["title","order","category","questions"]},"JobResponsibility":{"type":"string","enum":["INDIVIDUAL_CONTRIBUTOR","PEOPLE_LEADER","DONT_CARE"]},"RemoteAllowance":{"type":"string","enum":["NO_REMOTE","PARTIAL_WEEK_REMOTE_IN_CITY","FULLY_REMOTE_IN_COUNTRY","REMOTE_WORLDWIDE","DONT_CARE"]},"CompanyAudience":{"type":"string","enum":["B_TO_C","B_TO_B","DONT_CARE"]},"CompanyFunding":{"type":"string","enum":["SELF_FUNDED","VENTURE_CAPITAL","PUBLIC","DONT_CARE"]},"ApplicantRole":{"type":"string","enum":["RECRUITER","HIRING_MANAGER","DONT_CARE"]},"FilterSetting":{"type":"object","properties":{"askAboutSalary":{"type":"boolean"},"minimumSalary":{"type":"number"},"minimumTotalCompensation":{"type":"number"},"askAboutTotalCompensation":{"type":"boolean"},"mustIncludeSuper":{"type":"boolean"},"askAboutIncludeSuper":{"type":"boolean"},"mustProvidesSomeMedicalBenefit":{"type":"boolean"},"askAboutProvidesSomeMedicalBenefit":{"type":"boolean"},"primaryResponsibility":{"$ref":"#/components/schemas/JobResponsibility"},"askAboutPrimaryResponsibility":{"type":"boolean"},"askForRoleTitle":{"type":"boolean"},"remoteAllowance":{"$ref":"#/components/schemas/RemoteAllowance"},"askAboutRemoteAllowance":{"type":"boolean"},"askAboutRoleLocation":{"type":"boolean"},"askAboutTellMeWho":{"type":"boolean"},"askAboutDescribeAPieceOfWork":{"type":"boolean"},"askAboutWhyInternalExternal":{"type":"boolean"},"minimumLearningTimePercentPerMonth":{"type":"number"},"askAboutMinimumLearningTimePercentPerMonth":{"type":"boolean"},"askMoreAboutLearningTime":{"type":"boolean"},"askAboutOngoingChallenges":{"type":"boolean"},"askTechnicalPrincipals":{"type":"boolean"},"rejectIfOvertimeExpected":{"type":"boolean"},"askAboutOvertimeExpected":{"type":"boolean"},"mustBeFamilyFriendly":{"type":"boolean"},"askFamilyFriendlyPolicies":{"type":"boolean"},"askMoreDetailsAboutFamilyFriendlyEnvironment":{"type":"boolean"},"requireActiveDiversityProgram":{"type":"boolean"},"askAboutDiversityProgram":{"type":"boolean"},"askMoreAboutDiversity":{"type":"boolean"},"companyAudience":{"$ref":"#/components/schemas/CompanyAudience"},"askAboutCompanyAudience":{"type":"boolean"},"companyFunding":{"$ref":"#/components/schemas/CompanyFunding"},"askAboutCompanyFunding":{"type":"boolean"},"askAboutCompetitiveAdvantage":{"type":"boolean"},"rejectIfMustSignNonCompete":{"type":"boolean"},"askAboutNonCompete":{"type":"boolean"},"askAboutRoleOfSubmitter":{"type":"boolean"},"applicantRole":{"$ref":"#/components/schemas/ApplicantRole"},"askAboutNameOfCompany":{"type":"boolean"},"lodgementsAllowedPerDay":{"type":"number"},"allowUnlimitedSelfLodgements":{"type":"boolean"},"id":{"type":"number"},"createdDate":{"format":"date-time","type":"string"}},"required":["askAboutSalary","minimumSalary","minimumTotalCompensation","askAboutTotalCompensation","mustIncludeSuper","askAboutIncludeSuper","mustProvidesSomeMedicalBenefit","askAboutProvidesSomeMedicalBenefit","primaryResponsibility","askAboutPrimaryResponsibility","askForRoleTitle","remoteAllowance","askAboutRemoteAllowance","askAboutRoleLocation","askAboutTellMeWho","askAboutDescribeAPieceOfWork","askAboutWhyInternalExternal","minimumLearningTimePercentPerMonth","askAboutMinimumLearningTimePercentPerMonth","askMoreAboutLearningTime","askAboutOngoingChallenges","askTechnicalPrincipals","rejectIfOvertimeExpected","askAboutOvertimeExpected","mustBeFamilyFriendly","askFamilyFriendlyPolicies","askMoreDetailsAboutFamilyFriendlyEnvironment","requireActiveDiversityProgram","askAboutDiversityProgram","askMoreAboutDiversity","companyAudience","askAboutCompanyAudience","companyFunding","askAboutCompanyFunding","askAboutCompetitiveAdvantage","rejectIfMustSignNonCompete","askAboutNonCompete","askAboutRoleOfSubmitter","applicantRole","askAboutNameOfCompany","lodgementsAllowedPerDay","allowUnlimitedSelfLodgements","id","createdDate"]},"FilterSettingDto":{"type":"object","properties":{"askAboutSalary":{"type":"boolean"},"minimumSalary":{"type":"number"},"minimumTotalCompensation":{"type":"number"},"askAboutTotalCompensation":{"type":"boolean"},"mustIncludeSuper":{"type":"boolean"},"askAboutIncludeSuper":{"type":"boolean"},"mustProvidesSomeMedicalBenefit":{"type":"boolean"},"askAboutProvidesSomeMedicalBenefit":{"type":"boolean"},"primaryResponsibility":{"$ref":"#/components/schemas/JobResponsibility"},"askAboutPrimaryResponsibility":{"type":"boolean"},"remoteAllowance":{"$ref":"#/components/schemas/RemoteAllowance"},"askAboutRemoteAllowance":{"type":"boolean"},"askAboutRoleLocation":{"type":"boolean"},"askAboutTellMeWho":{"type":"boolean"},"askAboutDescribeAPieceOfWork":{"type":"boolean"},"askAboutWhyInternalExternal":{"type":"boolean"},"minimumLearningTimePercentPerMonth":{"type":"number"},"askAboutMinimumLearningTimePercentPerMonth":{"type":"boolean"},"askMoreAboutLearningTime":{"type":"boolean"},"askAboutOngoingChallenges":{"type":"boolean"},"askTechnicalPrincipals":{"type":"boolean"},"rejectIfOvertimeExpected":{"type":"boolean"},"askAboutOvertimeExpected":{"type":"boolean"},"mustBeFamilyFriendly":{"type":"boolean"},"askFamilyFriendlyPolicies":{"type":"boolean"},"askMoreDetailsAboutFamilyFriendlyEnvironment":{"type":"boolean"},"requireActiveDiversityProgram":{"type":"boolean"},"askAboutDiversityProgram":{"type":"boolean"},"askMoreAboutDiversity":{"type":"boolean"},"companyAudience":{"$ref":"#/components/schemas/CompanyAudience"},"askAboutCompanyAudience":{"type":"boolean"},"companyFunding":{"$ref":"#/components/schemas/CompanyFunding"},"askAboutCompanyFunding":{"type":"boolean"},"askAboutCompetitiveAdvantage":{"type":"boolean"},"rejectIfMustSignNonCompete":{"type":"boolean"},"askAboutNonCompete":{"type":"boolean"},"askAboutRoleOfSubmitter":{"type":"boolean"},"applicantRole":{"$ref":"#/components/schemas/ApplicantRole"},"askAboutNameOfCompany":{"type":"boolean"},"lodgementsAllowedPerDay":{"type":"number"},"allowUnlimitedSelfLodgements":{"type":"boolean"}},"required":["askAboutSalary","minimumSalary","minimumTotalCompensation","askAboutTotalCompensation","mustIncludeSuper","askAboutIncludeSuper","mustProvidesSomeMedicalBenefit","askAboutProvidesSomeMedicalBenefit","primaryResponsibility","askAboutPrimaryResponsibility","remoteAllowance","askAboutRemoteAllowance","askAboutRoleLocation","askAboutTellMeWho","askAboutDescribeAPieceOfWork","askAboutWhyInternalExternal","minimumLearningTimePercentPerMonth","askAboutMinimumLearningTimePercentPerMonth","askMoreAboutLearningTime","askAboutOngoingChallenges","askTechnicalPrincipals","rejectIfOvertimeExpected","askAboutOvertimeExpected","mustBeFamilyFriendly","askFamilyFriendlyPolicies","askMoreDetailsAboutFamilyFriendlyEnvironment","requireActiveDiversityProgram","askAboutDiversityProgram","askMoreAboutDiversity","companyAudience","askAboutCompanyAudience","companyFunding","askAboutCompanyFunding","askAboutCompetitiveAdvantage","rejectIfMustSignNonCompete","askAboutNonCompete","askAboutRoleOfSubmitter","applicantRole","askAboutNameOfCompany","lodgementsAllowedPerDay","allowUnlimitedSelfLodgements"]},"QuestionAnswer":{"type":"object","properties":{"raw":{"type":"object"},"display":{"type":"string"},"passedFilter":{"type":"boolean"}},"required":["raw","display"]},"QuestionResult":{"type":"object","properties":{"questionMeta":{"$ref":"#/components/schemas/QuestionMeta"},"filterValue":{"type":"object"},"answer":{"$ref":"#/components/schemas/QuestionAnswer"}},"required":["questionMeta","answer"]},"QuestionResultSection":{"type":"object","properties":{"title":{"type":"string"},"order":{"type":"number"},"category":{"$ref":"#/components/schemas/QuestionCategory"},"didPassAllFilters":{"type":"boolean"},"questionResults":{"type":"array","items":{"$ref":"#/components/schemas/QuestionResult"}}},"required":["title","order","category","didPassAllFilters","questionResults"]},"SubmittedOffer":{"type":"object","properties":{"applicantUuid":{"type":"string"},"offerUuid":{"type":"string"},"applicantName":{"type":"string"},"applicantEmail":{"type":"string"},"roleTitle":{"type":"string"},"questionResultSections":{"type":"array","items":{"$ref":"#/components/schemas/QuestionResultSection"}},"createdDate":{"format":"date-time","type":"string"},"id":{"type":"number"}},"required":["applicantUuid","offerUuid","applicantName","applicantEmail","roleTitle","questionResultSections","createdDate","id"]},"SharingLink":{"type":"object","properties":{"id":{"type":"number"},"ownerId":{"type":"string"},"uuid":{"type":"string"},"sharingMessage":{"type":"string"},"createdDate":{"format":"date-time","type":"string"},"updatedDate":{"format":"date-time","type":"string"},"deletedDate":{"format":"date-time","type":"string"}},"required":["id","ownerId","uuid","sharingMessage","createdDate","updatedDate","deletedDate"]},"OfferSubmission":{"type":"object","properties":{"offerUuid":{"type":"string"},"applicantName":{"type":"string"},"applicantRole":{"$ref":"#/components/schemas/ApplicantRole"},"applicantEmail":{"type":"string"},"roleTitle":{"type":"string"},"roleLocation":{"type":"string"},"salary":{"type":"number"},"totalCompensation":{"type":"number"},"inclusiveOfPension":{"type":"boolean"},"providesSomeMedicalBenefit":{"type":"boolean"},"primaryResponsibility":{"$ref":"#/components/schemas/JobResponsibility"},"remoteAllowance":{"$ref":"#/components/schemas/RemoteAllowance"},"tellMeAboutWhoWillIBeWorkingWith":{"type":"string"},"describeHowAPieceOfWorkGoesFromIdeaToProduct":{"type":"string"},"whyNotFilledInternally":{"type":"string"},"learningTimePercentPerMonth":{"type":"number"},"whatAreSomeOngoingChallengesYouHaveYetToResolve":{"type":"string"},"tellMeAboutTechnicalPrincipalsOrVision":{"type":"string"},"isFamilyFriendly":{"type":"boolean"},"familyFriendlyTellMeMore":{"type":"string"},"isOvertimeExpected":{"type":"boolean"},"hasActiveDiversityProgram":{"type":"boolean"},"tellMeMoreAboutDiversity":{"type":"string"},"companyAudience":{"$ref":"#/components/schemas/CompanyAudience"},"companyFunding":{"$ref":"#/components/schemas/CompanyFunding"},"companyName":{"type":"string"},"whatIsYourCompetitiveAdvantage":{"type":"string"},"nonCompeteAgreementRequired":{"type":"boolean"}},"required":["offerUuid","applicantName","applicantRole","roleTitle"]}}},"paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Application Support"]}},"/authorise":{"get":{"operationId":"AppController_getHelloAuthorized","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Application Support"],"security":[{"bearer":[]}]}},"/init-user":{"post":{"operationId":"AppController_initUser","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InitUserDto"}}}}},"tags":["Application Support"],"security":[{"bearer":[]}]}},"/person":{"post":{"operationId":"PersonController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePersonDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Person"}}}}},"tags":["persons"],"security":[{"bearer":[]}]},"get":{"operationId":"PersonController_findSelf","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Person"}}}}},"tags":["persons"],"security":[{"bearer":[]}]}},"/person/{id}":{"patch":{"operationId":"PersonController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePersonDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Person"}}}}},"tags":["persons"],"security":[{"bearer":[]}]},"delete":{"operationId":"PersonController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Person"}}}}},"tags":["persons"],"security":[{"bearer":[]}]}},"/filter-settings/latest":{"get":{"operationId":"FilterSettingsController_getLatest","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionSection"}}}}}},"tags":["filter-settings"],"security":[{"bearer":[]}]},"post":{"operationId":"FilterSettingsController_saveLatest","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilterSettingDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilterSetting"}}}}},"tags":["filter-settings"],"security":[{"bearer":[]}]}},"/filter-settings":{"get":{"operationId":"FilterSettingsController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilterSetting"}}}}}},"tags":["filter-settings"],"security":[{"bearer":[]}]},"delete":{"operationId":"FilterSettingsController_deleteAll","parameters":[],"responses":{"200":{"description":""}},"tags":["filter-settings"],"security":[{"bearer":[]}]}},"/filter-settings/settings/{offerId}":{"get":{"operationId":"FilterSettingsController_getSettings","parameters":[{"name":"offerId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionSection"}}}}}},"tags":["filter-settings"]}},"/received-roles":{"get":{"operationId":"ReceivedOffersController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SubmittedOffer"}}}}}},"tags":["received-roles"],"security":[{"bearer":[]}]},"delete":{"operationId":"ReceivedOffersController_deleteAll","parameters":[],"responses":{"200":{"description":""}},"tags":["received-roles"],"security":[{"bearer":[]}]}},"/sharing-links":{"post":{"operationId":"SharingLinksController_generateNew","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SharingLink"}}}}},"tags":["sharing-links"],"security":[{"bearer":[]}]},"get":{"operationId":"SharingLinksController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SharingLink"}}}}}},"tags":["sharing-links"],"security":[{"bearer":[]}]},"delete":{"operationId":"SharingLinksController_deleteAll","parameters":[],"responses":{"200":{"description":""}},"tags":["sharing-links"],"security":[{"bearer":[]}]}},"/sharing-links/latest":{"get":{"operationId":"SharingLinksController_getLatest","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SharingLink"}}}}},"tags":["sharing-links"],"security":[{"bearer":[]}]}},"/offer-links/validate-link":{"get":{"operationId":"OfferLinksController_validateLink","parameters":[{"name":"linkUuid","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["offer-links"]}},"/submitted-roles":{"post":{"operationId":"SubmittedOffersController_saveOffer","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfferSubmission"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmittedOffer"}}}}},"tags":["submitted-roles"],"security":[{"bearer":[]}]},"get":{"operationId":"SubmittedOffersController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SubmittedOffer"}}}}}},"tags":["submitted-roles"],"security":[{"bearer":[]}]}},"/email-client/verify":{"get":{"operationId":"EmailClientController_verify","parameters":[],"responses":{"200":{"description":""}},"tags":["email-client"],"security":[{"bearer":[]}]}}}}